fix: inline fsReader
